Implemented CSVReader and DataObjects for tasks
Implemented a CSVFileReader, which implements a DataFileReader interface. It reads a csv-file line by line, transforms each line into a data object and adds those to a ArrayList.
Created a generic DataObject interface and a more specific DataObjectValueDifference interface, which extends the DataObject interface. Implemented DataObjectWeather and DataObjectFootball and (for testing) DataObjectStandard as implementations of DataObjectValueDifference. DataObjectWeather has the attributes day, maxTemp and minTemp, DataObjectFootball has the attributes teamName, nrGoals and nrGoalsAllowed. Both implement a method getLabel, which provides the day or temName and a method getDifference, which provides the difference between maxTemp and minTemp or the absolute difference nrGoals and nrGoalsAllowed.
Implemented DataObjectFactory, which has the method createValueDifferenceObject for creating DataObjectWeather or DataObjectFootball (or DataObjectStandard) objects. This is used by the CSVFileReader for creating the DataObject objects.
Created a DefaultValues class to store default values like file names and column names.
Created CSVFileReaderTest for testing the file reader and DataObjectTest for testing the weather and football data objects.