Implemented Weather and Football tasks
Created a TaskRunner interface with findResult method and a TaskRunnerFactory interface with a createTask method.
Implemented FindSmallestTask, which implements TaskRunner. It has a list of data objects as attribute and the findResult method iterates over the list to find the object for which getDifference returns the smallest result.
Implemented FindSmallestTaskFactory, which has a DataFileReader and DataObjectType as class attributes. The file reader can either be provided on creation or it is set according to the provided dataObjectType and optionally a non-default file path. FindSmallestTaskFactory implements the method createTask, which creates a FindSmallestTask task with the data read via the file reader.
Created TaskRunnerTest for testing the TaskRunner implementation FindSmallestTask.
Did some slight refactoring and adjusted some comments.