Added command line argument handling
When users provide command line options, it is now handeled in the main method of App.java. If the args contain either "--weather" or "-football", the argument directly after is used as name of the input file.
Added tests for running the main method with different options.
Did some slight refactoring.